refactor: change enum package name
