+++ chat/src/main/kotlin/cobo/chat/config/jwt/JwtFilter.kt
@@ -1,6 +1,6 @@
 package cobo.chat.config.jwt
 
-import cobo.chat.data.enum.RoleEnum
+import cobo.chat.data.enums.RoleEnum
 import jakarta.servlet.FilterChain
 import jakarta.servlet.http.HttpServletRequest
 import jakarta.servlet.http.HttpServletResponse
