+++ chat/src/main/kotlin/cobo/chat/data/enums/RoleEnum.kt
@@ -1,4 +1,4 @@
-package cobo.chat.data.enum
+package cobo.chat.data.enums
 
 enum class RoleEnum(val value: Short) {
     DEVELOPER(0),
