+++ chat/src/main/kotlin/cobo/chat/repository/ChatRoomRepository.kt
@@ -1,7 +1,7 @@
 package cobo.chat.repository
 
 import cobo.chat.data.entity.ChatRoom
-import cobo.chat.data.enum.ChatStateEnum
+import cobo.chat.data.enums.ChatStateEnum
 import cobo.chat.repository.custom.ChatRoomRepositoryCustom
 import org.springframework.data.jpa.repository.JpaRepository
