+++ chat/src/main/kotlin/cobo/chat/repository/impl/ChatRoomRepositoryImpl.kt
@@ -2,7 +2,7 @@ package cobo.chat.repository.impl
 
 import cobo.chat.data.dto.prof.ProfGetListElementRes
 import cobo.chat.data.entity.ChatRoom
-import cobo.chat.data.enum.ChatStateEnum
+import cobo.chat.data.enums.ChatStateEnum
 import cobo.chat.repository.custom.ChatRoomRepositoryCustom
 import org.springframework.jdbc.core.JdbcTemplate
 import org.springframework.stereotype.Repository
