+++ chat/src/test/kotlin/cobo/chat/prof/ProfGetListTest.kt
@@ -2,7 +2,7 @@ package cobo.chat.prof
 
 import cobo.chat.data.entity.Chat
 import cobo.chat.data.entity.ChatRoom
-import cobo.chat.data.enum.ChatStateEnum
+import cobo.chat.data.enums.ChatStateEnum
 import cobo.chat.repository.ChatRepository
 import cobo.chat.repository.ChatRoomRepository
 import cobo.chat.service.ChatService
